pyproject.toml
[tool.poetry]
name = "qat-rpc"
version = "0.2.0"
description = "RPC tooling for OQC QAT."
authors = ["Kajsa Eriksson Rosenqvist <keriksson.rosenqvist@oxfordquantumcircuits.com>"]
readme = "README.md"
license = "BSD-3-Clause"
packages = [
    { include = "qat_rpc", from = "src/QAT_RPC/" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
qat-compiler = "^2.0.0"
pyzmq = "^26.1.0"
prometheus-client="^0.20.0"

[tool.poetry.group.dev.dependencies]
coverage = "^6.3.2"
pytest-cov = "~=2.10.1"
autoflake = "^2.2.0"
isort = "^5.13.2"
black = "24.3.0"
pre-commit = "^3.2.0"
pip-audit = "^2.4.5"
pip-licenses = "^5.0.0"

[tool.autoflake]
remove-all-unused-imports = true
ignore-init-module-imports = true
ignore-pass-after-docstring = true
in-place = true
recursive = true

[tool.isort]
py_version = 39
profile = "black"
line_length = 92
known_first_party = ["qat_rpc"]

[tool.black]
line-length = 92
target-version = ["py310", "py311", "py312"]

[tool.pip-licenses]
ignore-packages = [
  "docutils",
  "psycopg2",
  "qat-rpc"
]
fail-on = "GNU General Public License (GPL);GNU Library or Lesser General Public License (LGPL)"

[tool.poetry.scripts]
format-code = "poetry_scripts:format_code"
qat_comexe="qat_rpc.zmq.qat_commands:qat_run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
